% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2mzml-package.R
\docType{package}
\name{raw2mzml-package}
\alias{raw2mzml}
\alias{raw2mzml-package}
\title{raw2mzml: Convert Thermo Raw to mzML}
\description{
Shiny App to convert Thermo Raw to mzML using msConvert. This app only works on Windows OS.
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyun.joo@pennmedicine.upenn.edu}

}
